+++ test/eca/features/tools/filesystem_test.clj
@@ -22,7 +22,7 @@
     (is (match?
          {:contents [{:type :text
                       :error true
-                      :content "Access denied - path outside workspace root"}]}
+                      :content "Access denied - path outside workspace root, call list_allowed_dirs first"}]}
          ((get-in f.tools.filesystem/definitions ["list_directory" :handler])
           {"path" (h/file-path "/foo/qux")}
           {:workspace-folders [{:uri (h/file-uri "file:///foo/bar/baz") :name "baz"}]}))))
@@ -35,8 +35,8 @@
                                        (h/file-path "/foo/bar/baz/some.clj")
                                        (h/file-path "/foo/bar/baz/qux"))}]}
          (with-redefs [fs/starts-with? (constantly true)
-                       fs/list-dir (constantly [(fs/path "/foo/bar/baz/some.clj")
-                                                (fs/path "/foo/bar/baz/qux")])
+                       fs/list-dir (constantly [(fs/path (h/file-path "/foo/bar/baz/some.clj"))
+                                                (fs/path (h/file-path "/foo/bar/baz/qux"))])
                        fs/directory? (fn [path] (not (string/ends-with? (str path) ".clj")))]
            ((get-in f.tools.filesystem/definitions ["list_directory" :handler])
             {"path" (h/file-path "/foo/bar/baz")}